+++ zarrs/src/array/chunk_key_encoding.rs
@@ -1,6 +1,8 @@
 //! Zarr chunk key encodings. Includes a [default](default::DefaultChunkKeyEncoding) and [v2](v2::V2ChunkKeyEncoding) implementation.
 //!
 //! See <https://zarr-specs.readthedocs.io/en/latest/v3/chunk-key-encodings/index.html>.
+//!
+#![doc = include_str!("../../doc/status/chunk_key_encodings.md")]
 
 pub mod default;
 pub mod v2;